cloud-function/requirements.txt
functions-framework==3.8.2
google-cloud-storage==2.18.2